refactor:  models in different file, update .env
