+++ run.py
@@ -7,7 +7,6 @@ import dateutil.parser
 import babel
 from flask import Flask, render_template, request, Response, flash, redirect, url_for, abort, jsonify
 from flask_moment import Moment
-from flask_sqlalchemy import SQLAlchemy
 import logging
 from logging import Formatter, FileHandler
 from flask_wtf import form
@@ -39,6 +38,7 @@ from functools import wraps
 from jose import jwt
 from urllib.request import urlopen
 
+from models import db, Winemaker, Wine, Winery
 
 
 #----------------------------------------------------------------------------#
@@ -49,7 +49,7 @@ app = Flask(__name__)
 moment = Moment(app)
 app.config.from_object(env['APP_SETTINGS'])
 print(env['APP_SETTINGS'])
-db = SQLAlchemy(app)
+db.init_app(app)
 migrate = Migrate(app, db)
 
 # Auth0 login flow app config
@@ -70,80 +70,6 @@ auth0 = oauth.register(
 ALGORITHMS = ['RS256']
 API_AUDIENCE = 'wineport'
 AUTH0_DOMAIN = env['AUTH0_DOMAIN']
-
-#----------------------------------------------------------------------------#
-# Models.
-#----------------------------------------------------------------------------#
-
-class Winery(db.Model):
-    __tablename__ = 'winery'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String, unique=True)
-    city = db.Column(db.String(120))
-    state = db.Column(db.String(120))
-    address = db.Column(db.String(120))
-    phone = db.Column(db.String(120))
-    genres = db.Column(db.String(120))
-    facebook_link = db.Column(db.String(120))
-    image_link = db.Column(db.String(500))
-    website = db.Column(db.String(120))
-    seeking_talent = db.Column(db.Boolean, default=False)
-    seeking_description = db.Column(db.String(300))
-    wines = db.relationship('Wine', backref='winery', lazy='dynamic')
-
-    def __repr__(self):
-      return f'''< winery 
-                        id: {self.id},
-                      name: {self.name},
-                      city: {self.city},
-                     state: {self.state} >'''
-
-    # TODO: implement any missing fields, as a database migration using Flask-Migrate - done
-
-class Winemaker(db.Model):
-    __tablename__ = 'winemaker'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String, unique=True)
-    city = db.Column(db.String(120))
-    state = db.Column(db.String(120))
-    address = db.Column(db.String(120))
-    phone = db.Column(db.String(120))
-    genres = db.Column(db.String(120))
-    image_link = db.Column(db.String(500))
-    facebook_link = db.Column(db.String(120), nullable=True)
-    website = db.Column(db.String(120))
-    seeking_winery = db.Column(db.Boolean, default=False)
-    seeking_description = db.Column(db.String(300))
-    wines = db.relationship('Wine', backref='winemaker', lazy='dynamic')
-
-    def __repr__(self):
-      return f'''< winemaker 
-               id: {self.id},
-             name: {self.name},
-             city: {self.city},
-            state: {self.state}>'''
-
-    # TODO: implement any missing fields, as a database migration using Flask-Migrate - done
-
-class Wine(db.Model):
-    __tablename__ = 'wines'
-    id = db.Column(db.Integer, primary_key=True)
-    winery_id = db.Column(db.Integer, db.ForeignKey(
-        'winery.id'), nullable=False)
-    winemaker_id = db.Column(db.Integer, db.ForeignKey(
-        'winemaker.id'), nullable=False)
-    start_time = db.Column(db.DateTime(), nullable=False)
-
-    def __repr__(self):
-      return f'''\n
-            Wine: {self.id} 
-           Winery: {self.winery.name} 
-          Winemaker: {self.winemaker.name}
-      start_time: {self.start_time} '''
-
-
 
 #----------------------------------------------------------------------------#
 # Filters.
